Set `cudnn` executor tensors to query device (PR2120)
